+++ ParallaxGenLib/include/patchers/PatcherDefault.hpp
@@ -87,6 +87,8 @@ public:
   auto applyPatchSlots(const std::array<std::wstring, NUM_TEXTURE_SLOTS> &OldSlots,
                        const PatcherMatch &Match) -> std::array<std::wstring, NUM_TEXTURE_SLOTS> override;
 
+  void processNewTXSTRecord(const PatcherMatch &Match, const std::string &EDID = {}) override;
+
   /**
    * @brief Apply default shader to a shape (does nothing)
    *