+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - Removed simplicity of snow warning as the mod is not inherently incompatible
 - Fixed NULL output in ParallaxGen_Diff.json when PG patched nothing
 - PBR patcher clears hair_soft_lighting flag now
+- PBR texture swap JSONs are automatically generated
 
 ## [0.7.3] - 2024-12-09